+++ reapy/tools/__init__.py
@@ -1,8 +1,4 @@
 """Define tools such as Program and custom json module."""
 
 import reapy
-from .inside_reaper import InsideReaper
-if not reapy.is_inside_reaper():
-    from .dist_program import Program
-else:
-    from .program import Program
+from .inside_reaper import inside_reaper