+++ reapy/tools/json.py
@@ -2,6 +2,8 @@
 
 import importlib
 import json
+import operator
+import sys
 
 
 class ClassCache(dict):
@@ -25,6 +27,12 @@ class ReapyEncoder(json.JSONEncoder):
     def default(self, x):
         if hasattr(x, '_to_dict'):
             return x._to_dict()
+        elif callable(x):
+            return {
+                "__callable__": True,
+                "module_name": x.__module__,
+                "name": x.__qualname__
+            }
         return json.JSONEncoder.default(self, x)
 
 
@@ -37,7 +45,15 @@ def dumps(x):
 
 
 def object_hook(x):
-    if "__reapy__" not in x:
+    if "__reapy__" in x:
+        reapy_class = _CLASS_CACHE[x["class"]]
+        return reapy_class(*x["args"], **x["kwargs"])
+    elif "__callable__" in x:
+        module_name, name = x["module_name"], x["name"]
+        try:
+            module = sys.modules[module_name]
+        except KeyError:
+            module = importlib.import_module(module_name)
+        return operator.attrgetter(name)(sys.modules[module_name])
+    else:
         return x
-    reapy_class = _CLASS_CACHE[x["class"]]
-    return reapy_class(*x["args"], **x["kwargs"])